requirements.txt
# Python installation requirements for OMERO.qa
# =============================================
#
# This file is meant to be consumed by pip
#
#     pip install -r requirements.txt
#
Cython==0.16
Django==1.4.1
psycopg2==2.4.5
PIL==1.1.7
-e git+git://github.com/matplotlib/matplotlib.git#egg=matplotlib
-e hg+http://hg.saddi.com/flup-server#egg=flup
pygeoip==0.2.4
lxml==2.3.5
django-auth-ldap==1.1.1
ipaddr==2.1.10
